Better error page for not found files
Before, an intenal error with `git ls-tree fileName: not found` was shown as a generic error. This makes it distinguishable in the API and modifies the frontend a bit to special case and render a proper not found page.
